requirements.txt
numpy
pandas
pyarrow
scikit-learn
networkx
torch
pyyaml
matplotlib
jupyter
